+++ src/reactpy_django/static/reactpy_django/pyscript-custom.css
@@ -1,0 +1,3 @@
+py-script {
+    display: none;
+}
