+++ src/reactpy_django/static/reactpy_django/pyscript-hide-debug.css
@@ -1,0 +1,3 @@
+.py-error {
+    display: none;
+}
