+++ src/reactpy_django/templates/reactpy/pyscript_component.html
@@ -1,4 +1,4 @@
-{% if reactpy_class %}<div id="pyscript-{{reactpy_uuid}}" class="{{reactpy_class}}">{{reactpy_initial_html}}</div>
+{% if reactpy_class %}<div id="pyscript-{{pyscript_uuid}}" class="{{pyscript_class}}">{{pyscript_initial_html}}</div>
 {% endif %}
-{% if not reactpy_class %}<div id="pyscript-{{reactpy_uuid}}">{{reactpy_initial_html}}</div>{% endif %}
-<py-script async config='{{ reactpy_config }}'>{{ reactpy_executor }}</py-script>
+{% if not reactpy_class %}<div id="pyscript-{{pyscript_uuid}}">{{pyscript_initial_html}}</div>{% endif %}
+<py-script async>{{pyscript_executor}}</py-script>
