+++ tests/test_app/templates/pyscript.html
@@ -8,7 +8,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link rel="shortcut icon" type="image/png" href="{% static 'favicon.ico' %}" />
     <title>ReactPy</title>
-    {% pyscript_static_files %}
+    {% pyscript_setup %}
 </head>
 
 <body>
